fix: fixed enums naming and usage, made endpoint to retrive all possible enums for the frontend

+++ PharmaCore.Application/Auth/Services/GetCurrentUserService.cs
@@ -35,7 +35,7 @@ public class GetCurrentUserService : IGetCurrentUserService
                     user.UserName,
                     user.PhoneNumber,
                     user.Address,
-                    (short)user.Role,
+                    user.Role,
                     user.CreatedAt));
         }
         catch (Exception e)
